build/ops: in jessie virtualenv is in package virtualenv
The python scripts are not yet compatible with python3, yet it is the
default on jessie. Force the creation of the virtualenv to use python2.7
instead. The wheelhouse is already explicitly populated for both python3
and python2.7 by install-deps.sh, regardless of the default interpreter.

Signed-off-by: Loic Dachary <loic@dachary.org>
(cherry picked from commit eb583cbbbf1ec3e0abc730130afaf084ed9d0828)

Conflicts:
	qa/workunits/ceph-helpers.sh - removed, we only care about /debian/