rgw: resolve broken listing
Bucket listing filtering was moved from the RGW lyer to the CLS layer
to improve efficiency. However currently if there are enough entries
that are filtered out, the CLS call may return zero entries back to
RGW. Since we did not mark how far we got, calling it again will yield
the same result, causing the process to be stuck and fail.

This solution adds a marker to the CLS call's return object to
RGW. That will allow the next call to pick up where it left off.

Because the bucket index is spread across many shards, the
CLSRGWConcurrentIO class is used to coordinate these asynchronous
calls. Functionality is added to this class to handle re-issuing the
call with the new marker to make sure at least one entry is returned.

Additional clean-up is also done.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>